Profile-Card
I have build a simple profile card which is fully responsive. I have completed this project using only html and css. The special thing of this project is that it has responsive image, responsive font and all other elements are responsive and also media queries are not used in this project.